refactor: 컨트롤러 레이어 삭제 및 관련 클래스 이름 변경 개선
